Movement of the snake
The snake now can move using the keyboard (left, right, up and down).

On the other hand, is not program to eat the apple (red).

And the snake can move in space and doesn't matter if the snake goes out from the space created (black-background) because it can come back using the keyboard to return the snake.